More unit testing
I noticed a few things are in a strange order. I am making sure that the only new code that a test is running is the new code it is dedicated to testing.

Also, added a bit of type-testing. Arrays are weird, so they have a dedicated function, and show up as two tests.